app_webhook/requirements/base.txt
fastapi==0.112.2
pydantic-settings==2.4.0
uvicorn==0.30.6
pika==1.3.2